host-build.mk: avoid installing $(HOST_BUILD_PREFIX)/share/info/dir
The problem is that this file is essentially a listing of all installed
info pages. This means when we assemble temporary build dirs with just
the dependencies of a particular tool, this file would be typically
incorrect and lead to conflicting $(HOST_BUILD_PREFIX)/share/info/dir
files in different tools. This breaks checksumming all files, which is
done in our current approach.